+++ app/Http/Controllers/InstallController.php
@@ -50,7 +50,7 @@ class InstallController extends Controller
                 $table->boolean("switch");
                 $table->text("reason");
                 $table->enum("prov", ["北京市", "天津市", "上海市", "重庆市", "河北省", "山西省", "内蒙古自治区", "辽宁省", "吉林省", "黑龙江省", "江苏省", "浙江省", "安徽省", "福建省", "江西省", "山东省", "河南省", "湖北省", "湖南省", "广东省", "广西壮族自治区", "海南省", "四川省", "贵州省", "云南省", "西藏自治区", "陕西省", "甘肃省", "青海省", "宁夏回族自治区", "新疆维吾尔自治区", "香港特别行政区", "澳门特别行政区", "台湾省"])->nullable();
-                $table->timestamp("used_at");
+                $table->dateTime("used_at");
                 $table->timestamps();
             });
 
@@ -60,7 +60,7 @@ class InstallController extends Controller
                 $table->enum("type", ["ip", "fingerprint"]);
                 $table->text("identifier");
                 $table->text("reason");
-                $table->timestamp("expires_at");
+                $table->dateTime("expires_at");
                 $table->timestamps();
             });
 
@@ -97,7 +97,9 @@ class InstallController extends Controller
                 $table->unsignedBigInteger("day");
                 $table->unsignedBigInteger("can_use_ip_count");
                 $table->json("ip");
-                $table->timestamp("expires_at")->nullable();
+                $table->boolean("switch");
+                $table->text("reason");
+                $table->dateTime("expires_at")->nullable();
                 $table->timestamps();
             });
 
@@ -109,6 +111,8 @@ class InstallController extends Controller
                 "day" => 0,
                 "can_use_ip_count" => 0,
                 "ip" => [],
+                "switch" => true,
+                "reason" => "",
                 "expires_at" => "2099-01-01 00:00:00",
             ]);
 
@@ -119,7 +123,7 @@ class InstallController extends Controller
         }
 
         UtilsController::updateEnv([
-            'APP_NAME' => '"' . $request['app_name'] . '"',
+            'APP_NAME' => $request['app_name'],
             'DB_CONNECTION' => "mysql",
             'DB_HOST' => $request['db_host'],
             'DB_PORT' => $request['db_port'],